Update mods
+ Quark Oddities
+ Immersive Weathering
+ Carry On
+ Vertical Slabs Compat
+ Gravestone
+ Creeper Firework
+ Simple Backups
~ Disabled Quark's variant ladder and chests
^ AlmostUnified
^ Amendments
^ Architectury
^ Balm
^ Citadel
^ Crate Delight
^ Delightful
^ Kiwi
^ Quark
^ Sawmill
^ Small Ships
^ Supplementaries
^ Titanium
^ YUNG's API
^ Zeta
- Crafting Tweaks

+++ build.ps1
@@ -91,7 +91,9 @@ rg '<!-- ([A-Z]+) -->' .\src\template\ --files-with-matches | ForEach-Object {
 }
 
 $ArtifactName = "$($Manifest.name)-v$($Manifest.version)".Replace(' ', '-')
-Remove-Item -Force "$(Resolve-Path .\build\artifact\)$ArtifactName.zip"
+if (Test-Path "$(Resolve-Path .\build\artifact\)$ArtifactName.zip") {
+    Remove-Item -Force "$(Resolve-Path .\build\artifact\)$ArtifactName.zip"
+}
 7z u "$(Resolve-Path .\build\artifact\)$ArtifactName.zip" "$(Resolve-Path .\pack\)*" | Out-Null
 Write-Host "$(TimeCode) Created artifact: $(Resolve-Path .\build\artifact\)$ArtifactName.zip"
